Added a task status named "finished" that will be set when the task has finished rendering but hasn't yet finished sending the update to the master. That task status fixes the issue of having messages pointing out that there should be a task which couldn't be found (the pid disappeared, but the task status still said running).
More debug messages on the new log levels L_DEBUG2 and L_DEBUG3.
Log messages written to the task log, are now also copied to the slave log too.
Send the computer status with the newer send_computer_loadavg and send_computer_ntasks that were not being used before.
The last connection time updating code has been moved before the "switch" because not all master request handlers return, some of them cause the process to die or exit. And then the time was not updated properly.

git-svn-id: https://ssl.drqueue.org/svn/branches/0.64.x@1561 94d6b6b3-0ee6-0310-ab95-9b4161196ee2